Split translate into translate and tools
- translate is for simple translators
- tools is for compound tools that might involve some translations
