+++ pgmpl/contour.py
@@ -23,7 +23,8 @@ from pyqtgraph import functions as fn
 # pgmpl
 # noinspection PyUnresolvedReferences
 import pgmpl.__init__  # __init__ does setup stuff like making sure a QApp exists
-from pgmpl.translate import setup_pen_kw, color_map_translator
+from pgmpl.translate import color_map_translator
+from pgmpl.tools import setup_pen_kw
 from pgmpl.util import printd, tolist
